Delete 11 test files for extracted modules, add parity tests, update docs and dependencies
Delete unit tests for modules moved to wxyc-etl and wxyc-catalog: test_matching, test_artist_splitting, test_catalog_source, test_wxyc, test_export, test_enrich_library_artists, test_extract_library_labels, test_pipeline_state. Delete integration tests: test_catalog_source, test_db_introspect, test_enrich_library_artists. Add test_wxyc_etl_parity.py with 30 characterization tests verifying behavioral parity with the new shared packages. Update pyproject.toml to declare wxyc-etl and wxyc-catalog dependencies, remove pymysql/mysql optional deps. Fix run_pipeline.py extract_library_labels subprocess call to use wxyc-catalog CLI. Update sync-library.sh and sync-library.yml to use wxyc-export-to-sqlite. Update Dockerfile, CLAUDE.md, and README.md.
