+++ CLAUDE.md
@@ -12,7 +12,7 @@ ETL pipeline for building and maintaining a PostgreSQL cache of Discogs release 
 ### Pipeline Steps
 
 1. **Download** Discogs monthly data dumps (XML) from https://discogs-data-dumps.s3.us-west-2.amazonaws.com/index.html
-2. **Enrich** `library_artists.txt` with WXYC cross-references (`scripts/enrich_library_artists.py`, optional)
+2. **Enrich** `library_artists.txt` with WXYC cross-references (via `wxyc-enrich-library-artists` CLI from wxyc-catalog, optional)
 3. **Convert and filter** XML to CSV using [discogs-xml-converter](https://github.com/WXYC/discogs-xml-converter) (Rust binary), with optional artist filtering via `--library-artists`. Accepts a single XML file or a directory containing releases.xml, artists.xml, and labels.xml. When artists.xml is present, alias-enhanced filtering is enabled automatically. When labels.xml is present, `label_hierarchy.csv` is produced for sublabel-aware dedup.
 4. **Create schema** (`schema/create_database.sql`) and **functions** (`schema/create_functions.sql`), then **SET UNLOGGED** on all tables to skip WAL writes during bulk import
 5. **Import** filtered CSVs into PostgreSQL (`scripts/import_csv.py`)
@@ -66,7 +66,6 @@ docker compose up db -d     # just the database (for tests)
 ### Key Files
 
 - `scripts/run_pipeline.py` -- Pipeline orchestrator (--xml for steps 2-9, --csv-dir for steps 4-9)
-- `scripts/enrich_library_artists.py` -- Enrich artist list with WXYC cross-references via CatalogSource
 - `scripts/filter_csv.py` -- Filter Discogs CSVs to library artists (standalone, used outside the pipeline)
 - `scripts/import_csv.py` -- Import CSVs into PostgreSQL (psycopg COPY). Child tables are imported in parallel via ThreadPoolExecutor after parent tables. Artist detail tables (artist_alias, artist_member) are filtered to known artist IDs to prevent FK violations, since the converter's CSVs contain all Discogs artists. Tables with `unique_key` configs are deduped in-memory during COPY.
 - `scripts/dedup_releases.py` -- Deduplicate releases by master_id, preferring label match + sublabel resolution, US releases (copy-swap with `DROP CASCADE`). Index/constraint creation is parallelized via ThreadPoolExecutor.
@@ -74,25 +73,25 @@ docker compose up db -d     # just the database (for tests)
 - `scripts/csv_to_tsv.py` -- CSV to TSV conversion utility
 - `scripts/fix_csv_newlines.py` -- Fix multiline CSV fields
 - `lib/format_normalization.py` -- Normalize raw Discogs/library format strings to broad categories (Vinyl, CD, Cassette, 7", Digital)
-- `lib/artist_splitting.py` -- Split combined multi-artist library entries into individual components for matching
-- `lib/catalog_source.py` -- CatalogSource protocol with TubafrenzySource (MySQL) and BackendServiceSource (PostgreSQL) implementations for querying the WXYC library catalog
-- `lib/matching.py` -- Compilation detection utility
-- `lib/pipeline_state.py` -- Pipeline state tracking for resumable runs
-- `lib/db_introspect.py` -- Database introspection for inferring pipeline state on resume
-- `scripts/export_to_sqlite.py` -- Generate `library.db` from WXYC catalog via SSH (default) or `--catalog-source`
-- `scripts/extract_library_labels.py` -- Extract label preferences from flowsheet plays via CatalogSource
 - `scripts/sync-library.sh` -- Orchestrate library.db generation + upload to library-metadata-lookup
 - `docs/discogs-cache-technical-overview.md` -- Design rationale, benchmarks, and pipeline architecture details
+
+### Shared Package Dependencies
+
+Functionality that was previously local to this repo has been extracted to shared packages:
+
+- **wxyc-etl** (Rust/PyO3) -- Artist name normalization (`normalize_artist_name`), compilation detection (`is_compilation_artist`), artist name splitting (`split_artist_name`, `split_artist_name_contextual`), pipeline state tracking (`PipelineState`), and database introspection.
+- **wxyc-catalog** -- Catalog source protocol (`CatalogSource`, `TubafrenzySource`, `BackendServiceSource`), library.db export (`wxyc-export-to-sqlite` CLI), library artist enrichment (`wxyc-enrich-library-artists` CLI), and label extraction (`wxyc-extract-library-labels` CLI).
 
 ### External Inputs
 
 ### Library Catalog (library.db)
 
-`library.db` is a SQLite export of the WXYC library catalog, generated by `scripts/export_to_sqlite.py`. The default source is tubafrenzy's MySQL database on Kattare (via SSH), but Backend-Service's PostgreSQL database can be used instead via `--catalog-source backend-service --catalog-db-url <url>`. The `--catalog-source` flag is supported by `export_to_sqlite.py`, `enrich_library_artists.py`, `extract_library_labels.py`, and `run_pipeline.py`. The legacy `--wxyc-db-url` flag is an alias for `--catalog-source tubafrenzy --catalog-db-url <url>`. Backend-Service as a catalog source requires migration 0031 (`artist_crossreference` table) from Backend-Service PR #265.
+`library.db` is a SQLite export of the WXYC library catalog, generated by the `wxyc-export-to-sqlite` CLI from the wxyc-catalog package. The default source is tubafrenzy's MySQL database on Kattare (via SSH), but Backend-Service's PostgreSQL database can be used instead via `--catalog-source backend-service --catalog-db-url <url>`. The `--catalog-source` flag is supported by `run_pipeline.py` and the wxyc-catalog CLI tools. The legacy `--wxyc-db-url` flag is an alias for `--catalog-source tubafrenzy --catalog-db-url <url>`.
 
 `library.db` is used as input throughout the pipeline:
 
-1. **`library_artists.txt`** -- Generated from `library.db` by `scripts/enrich_library_artists.py`, one artist name per line, used by `discogs-xml-converter --library-artists` for filtering
+1. **`library_artists.txt`** -- Generated from `library.db` by `wxyc-enrich-library-artists` (wxyc-catalog CLI), one artist name per line, used by `discogs-xml-converter --library-artists` for filtering
 2. **KEEP/PRUNE classification** -- `scripts/verify_cache.py` uses `library.db` to match cached releases against the WXYC catalog
 
 `scripts/sync-library.sh` orchestrates the full flow: generate `library.db`, then upload it to library-metadata-lookup staging and production. The pipeline can also generate `library.db` inline via `--generate-library-db`.
@@ -111,15 +110,12 @@ pytest tests/unit/ -v
 DATABASE_URL_TEST=postgresql://discogs:discogs@localhost:5433/postgres \
   pytest -m postgres -v
 
-# MySQL integration tests (needs WXYC MySQL on port 3307)
-pytest -m mysql -v
-
 # E2E tests (runs full pipeline as subprocess against test Postgres)
 DATABASE_URL_TEST=postgresql://discogs:discogs@localhost:5433/postgres \
   pytest -m e2e -v
 ```
 
-Markers: `postgres` (needs PostgreSQL), `mysql` (needs WXYC MySQL), `e2e` (full pipeline), `integration` (needs library.db). Integration and E2E tests are excluded from the default `pytest` run via `addopts` in `pyproject.toml`.
+Markers: `postgres` (needs PostgreSQL), `e2e` (full pipeline), `integration` (needs library.db). Integration and E2E tests are excluded from the default `pytest` run via `addopts` in `pyproject.toml`.
 
 Test fixtures are in `tests/fixtures/` (CSV files, library.db, library_artists.txt). Regenerate with `python tests/fixtures/create_fixtures.py`.
 
@@ -137,7 +133,7 @@ This pipeline runs monthly (or when Discogs publishes new data dumps). It has a 
 
 ### Library Sync (`sync-library.yml`)
 
-A GitHub Actions cron workflow runs `scripts/sync-library.sh` daily at noon UTC (7 AM EST / 8 AM EDT) to export the WXYC library catalog from MySQL on Kattare to SQLite and upload it to library-metadata-lookup staging and production environments.
+A GitHub Actions cron workflow runs `scripts/sync-library.sh` daily at noon UTC (7 AM EST / 8 AM EDT) to export the WXYC library catalog to SQLite (via `wxyc-export-to-sqlite` from wxyc-catalog) and upload it to library-metadata-lookup staging and production environments.
 
 The workflow can also be triggered manually: `gh workflow run sync-library.yml`
