+++ README.md
@@ -16,7 +16,9 @@ The pipeline processes monthly Discogs data dumps (~63.3 GB XML) into a focused 
 - PostgreSQL with the `pg_trgm` and `unaccent` extensions (or use [Docker Compose](#docker-compose))
 - [discogs-xml-converter](https://github.com/WXYC/discogs-xml-converter) -- Rust binary for XML-to-CSV conversion (build from source or install on PATH)
 - Discogs monthly data dumps (XML) from https://discogs-data-dumps.s3.us-west-2.amazonaws.com/index.html -- `releases.xml.gz` is required; `artists.xml.gz` and `labels.xml.gz` are optional for enhanced filtering and dedup
-- `library.db` -- SQLite export of the WXYC library catalog, generated by `scripts/export_to_sqlite.py` (requires SSH tunnel to WXYC MySQL)
+- `library.db` -- SQLite export of the WXYC library catalog, generated by the `wxyc-export-to-sqlite` CLI from the [wxyc-catalog](https://github.com/WXYC/wxyc-catalog) package
+- [wxyc-etl](https://github.com/WXYC/wxyc-etl) -- Rust/PyO3 package providing artist name normalization, compilation detection, artist splitting, and pipeline state tracking
+- [wxyc-catalog](https://github.com/WXYC/wxyc-catalog) -- Python package providing catalog source access, library.db export, artist enrichment, and label extraction CLIs
 
 ## Quick Start
 
@@ -24,10 +26,10 @@ The pipeline processes monthly Discogs data dumps (~63.3 GB XML) into a focused 
 uv sync
 ```
 
-Generate `library.db` (requires SSH tunnel to WXYC MySQL):
+Generate `library.db` (requires SSH tunnel to WXYC MySQL, uses wxyc-catalog CLI):
 
 ```bash
-python scripts/export_to_sqlite.py   # see script for required env vars
+wxyc-export-to-sqlite   # see wxyc-catalog docs for required env vars
 ```
 
 Run the pipeline:
@@ -49,7 +51,7 @@ All 8 steps are automated by `run_pipeline.py`. The script supports two modes: f
 
 | Step | Script | Description |
 |------|--------|-------------|
-| 1. Generate artist list | `scripts/enrich_library_artists.py` | Extract artists from library.db, optionally enrich with cross-references |
+| 1. Generate artist list | `wxyc-enrich-library-artists` (wxyc-catalog) | Extract artists from library.db, optionally enrich with cross-references |
 | 2. Convert + Filter | `discogs-xml-converter` | XML to CSV with optional artist filtering (alias-enhanced when artists.xml is present) |
 | 3. Create schema | `schema/create_database.sql`, `schema/create_functions.sql` | Set up tables, extensions, and functions |
 | 4. Import | `scripts/import_csv.py` | Bulk load CSVs via psycopg COPY |
@@ -113,13 +115,13 @@ python scripts/run_pipeline.py \
 
 By default, dedup keeps the release with the most tracks per `master_id` group. When WXYC label preferences are available, dedup instead prefers the release whose Discogs label matches WXYC's known pressing -- ensuring the cached edition matches what the station actually owns.
 
-Label preferences come from WXYC's `FLOWSHEET_ENTRY_PROD` MySQL table (rotation play entries include `LABEL_NAME`). The extraction script `scripts/extract_library_labels.py` produces a `library_labels.csv` with `(artist_name, release_title, label_name)` triples.
+Label preferences come from the WXYC catalog (rotation play entries include `LABEL_NAME`). The extraction CLI `wxyc-extract-library-labels` (from wxyc-catalog) produces a `library_labels.csv` with `(artist_name, release_title, label_name)` triples.
 
 There are two ways to enable label-aware dedup:
 
-1. **Automatic extraction** (when `--wxyc-db-url` is provided): the pipeline extracts labels from MySQL before the dedup step.
+1. **Automatic extraction** (when `--wxyc-db-url` or `--catalog-source` is provided): the pipeline extracts labels via the wxyc-catalog CLI before the dedup step.
 
-2. **Pre-generated CSV** (when `--library-labels` is provided): the pipeline uses the CSV directly, no MySQL connection needed.
+2. **Pre-generated CSV** (when `--library-labels` is provided): the pipeline uses the CSV directly, no catalog connection needed.
 
 ```bash
 # Automatic: extract labels from WXYC MySQL and use for dedup
@@ -322,9 +324,6 @@ pytest tests/unit/ -v
 # Integration tests (needs PostgreSQL)
 DATABASE_URL_TEST=postgresql://discogs:discogs@localhost:5433/postgres \
   pytest -m postgres -v
-
-# MySQL integration tests (needs WXYC MySQL on port 3307)
-pytest -m mysql -v
 
 # E2E tests (needs PostgreSQL, runs full pipeline as subprocess)
 DATABASE_URL_TEST=postgresql://discogs:discogs@localhost:5433/postgres \
